% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_part_set.R
\name{align_part_set}
\alias{align_part_set}
\title{Co-optimal alignments from tree partitions}
\usage{
align_part_set(x, partition_set, msa.exec, method, coopt.sub = "all")
}
\arguments{
\item{x}{An object of class \code{\link{DNAbin}} or
\code{\link{AAbin}} containing unaligned sequences of DNA or amino acids. Here inherited from function \code{\link{guidance2}} or \code{\link{HoT}}.}

\item{partition_set}{partitions of the alignment based on the guide tree of the base MSA. Here inherited from the function \code{\link{partitions}}.}

\item{msa.exec}{A character string giving the path to the executable of the
alignment program (e.g. \code{/usr/local/bin/mafft}); possible programs are
\code{MAFFT}, \code{MUSCLE}, and \code{ClustalW}.}

\item{method}{A character string containing further arguments passed to
MAFFT; default is \code{"auto"}.}

\item{coopt.sub}{option for selection of specific co-optimal alignments, currently not available and set to "all"}
}
\description{
Produce co-optimal MSAs from the N-3 bipartitions of the
  starting tree. This function performs one iteration i 1: Take i partition
  and align both sides Heads (H) and Tails (T) and we get H1,H2 and T1, T2 2:
  align combinations again Heads and Tails: H1H2, H1T1, H1T2, T1T2 => 8
  combinations
}
\author{
Franz-Sebastian Krah
}
