% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMSA.R
\name{filterMSA}
\alias{filterMSA}
\title{Filtering of the base MSA using guidance scores}
\usage{
filterMSA(guidanceX, col.cutoff = 0.2, seq.cutoff = 0.1,
  mask.cutoff = 0.5, filter.ends = FALSE, filter.gaps = FALSE,
  na.coding = 0.5)
}
\arguments{
\item{guidanceX}{object of class \code{\link{guidance}}}

\item{col.cutoff}{numeric between 0 and 1; removes unreliable columns below
the cutoff (default: 0.2); ignored if FALSE}

\item{seq.cutoff}{numeric between 0 and 1; removes unreliable sequences below
the cutoff (default: 0.1); ignored if FALSE}

\item{mask.cutoff}{residues below the cutoff are masked ('N' for DNA, 'X' for
AA; default: 0.5); ignored if FALSE}

\item{filter.ends}{logical, if TRUE trim.ends (ips) is applied to the MSA}

\item{filter.gaps}{logical, if TRUE trim.gabs (ips) is applied to the MSA}

\item{na.coding}{value with which to replace NAs}
}
\value{
filtered MSA of class \code{AAbin} or \code{DNAbin}

if filter.ends or filter.gaps was choosen, the adjusted
}
\description{
Filtering of the base MSA using guidance scores
}
\details{
Please note that we do not recommend filtering MSAs (Tan et al. 2015). Instead the user should use the 
GUIDANCE column score as individual weights to each column of the alignment to RAxMLvia the –a flag 
(for a practical example: Krah et al. (2018)). This can be achieved by using the function scores with score = "column_raxml"
and passing the score as weights to \code{\link{raxml}}. For a detailled example, see Vignette.
}
\references{
Tan et al. (2015). Current methods for automated filtering of
  multiple sequence alignments frequently worsen single-gene phylogenetic
  inference. \emph{Systematic biology} \strong{64}:778--791.

Krah et al. (2018). Evolutionary dynamics of host specialization in wood-decay fungi. 
\emph{BMC Evolutionary Biology} \strong{18}:119-132
}
\seealso{
\code{\link{scores}}
}
\author{
Franz-Sebastian Krah
}
