% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidance.R
\name{guidance}
\alias{guidance}
\title{MSA Reliability Assessment with GUIDANCE}
\usage{
guidance(sequences, bootstrap = 100, method = "auto",
  msa.exec = "/usr/local/bin/mafft", ncore = 1, zip.file)
}
\arguments{
\item{sequences}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}} containing unaligned sequences of DNA or amino acids.}

\item{bootstrap}{An integer giving the number of alternative MSAs to be computed.}

\item{method}{A character string containing further arguments passed to MAFFT; default is \code{"auto"}.}

\item{msa.exec}{A character string giving the path to the executable of the alignment program (e.g. \code{/usr/local/bin/mafft}); possible programs are \code{MAFFT}, \code{MUSCLE}, \code{ClustalO}, and \code{ClustalW}.}

\item{ncore}{An integer specifying the number of cores; default = 1 (i.e. serial execution); \code{"auto"} can be used for automated usage of all detected cores.}

\item{zip.file}{A character string giving the dir of zip-compressed file to be produced, which contains the  alternative MSAs. If left empty (default), the alternative MSA will not be stored and cannot be assessed by the user.}
}
\value{
An object of class \code{\linkS4class{guidanceDNA}} or \code{\linkS4class{guidanceAA}}.
}
\description{
Calculate MSA reliability scores with GUIDANCE (Penn et al. 2010).
}
\details{
Calculates column confidence (and other scors) by comparing
  alternative MSAs generated by alternative guide trees derived from
  bootstrap MSAs (Felsenstein 1985). The basic comparison between the BP MSAs
  and a reference MSA is if column residue pairs are identically aligned in
  all alternative MSAs compared with the base MSA (see \code{msa_set_score}).
}
\examples{
\dontrun{
# run GUIDANCE on example data using MAFFT
file <- system.file("extdata", "BB50009.fasta", package = "polenta")
aa_seq <- read.fas(file)
g_res <- guidance(sequences = aa_seq)
scores <- daughter_scores(g_res, score = c("gcsc", "rprsc"))
hist(scores$gcsc$score, xlab = "Column score", main = "GUIDANCE")
}

}
\references{
Felsenstein, J. 1985. Confidence limits on phylogenies: an
  approach using the bootstrap. \emph{Evolution} \strong{39}:783--791.

Landan, G. and D. Graur. 2008. Local reliability measures from
  sets of co-optimal multiple sequence alignments. \emph{Pacific Symposium on
  Biocomputing} \strong{13}:15--24.

Penn, O., E. Privman, G. Landan, D. Graur, and T. Pupko. 2010. An
  alignment confidence score capturing robustness to guide tree uncertainty.
  \emph{Molecular Biology and Evolution} \strong{27}:1759--1767.
}
\seealso{
\code{\link{msa_set_score}}, \code{\link{guidance2}}, \code{\link{HoT}}
}
\author{
Franz-Sebastian Krah
}
