% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidance2.R
\name{guidance2}
\alias{guidance2}
\title{MSA Reliability Assessment with GUIDANCE2}
\usage{
guidance2(sequences, bootstrap = 100, method = "auto", n.coopt = 4,
  msa.exec = "/usr/local/bin/mafft", ncore = 1, zip.file)
}
\arguments{
\item{sequences}{An object of class \code{\link{DNAbin}} or
\code{\link{AAbin}} containing unaligned sequences of DNA or amino acids.}

\item{bootstrap}{An integer giving the number of alternative MSAs to be
computed.}

\item{method}{A character string containing further arguments passed to
MAFFT; default is \code{"auto"}.}

\item{n.coopt}{An integer giving the number of sampled co-optimal MSAa.}

\item{msa.exec}{A character string giving the path to the executable of the
alignment program (e.g. \code{/usr/local/bin/mafft}); possible programs are
\code{MAFFT}, \code{MUSCLE}, and \code{ClustalW}. 
For details see \code{\link{clustal}}, \code{\link{mafft}}}

\item{ncore}{An integer specifying the number of cores; default = 1 (i.e.
serial execution); \code{"auto"} can be used for automated usage of all
detected cores.}

\item{zip.file}{A character string giving the name of zip-compressed file,
which contains the alternative MSAs. If left empty (default), the
alternative MSA will not be stored and cannot be assessed by the user.}
}
\value{
An object of class \code{\linkS4class{guidanceDNA}} or
  \code{\linkS4class{guidanceAA}}.
}
\description{
Calculate MSA reliability scores with GUIDANCE2 (Sela et al.
  2015).
}
\details{
Calculates column confidence (and other scors) by comparing
  alternative MSAs generated by the GUIDANCE with varying gap opening panelty
  and the HoT methodology. First 100 alternative MSAs (with BP guide trees)
  with varying gap opening panelty are produced, then for each n (default =
  4) co-optimal alignments are produced using HoT. The basic comparison
  between the BP MSAs and a reference MSA is if column residue pairs are
  identically aligned in all alternative MSAs compared with the base MSA (see
  \code{compareMSAs}).

For an example workflow see Vignette
}
\examples{
\dontrun{
# run GUIDANCE on example data using MAFFT
fpath <- system.file("extdata", "BB30015.fasta", package="rGUIDANCE") # random example from BALiBASE
fas <- ape::read.FASTA(fpath)
g <- guidance2(sequences = fas, msa.exec= "/usr/local/bin/mafft")
scores <- scores(g, score = "column")
plot(scores$column$score, xlab = "Site", 
ylab = "Column score", 
main = "GUIDANCE2", type ="l")
}
}
\references{
Felsenstein, J. 1985. Confidence limits on phylogenies: an
  approach using the bootstrap. \emph{Evolution} \strong{39}:783--791.

Landan and Graur. 2008. Local reliability measures from
  sets of co-optimal multiple sequence alignments. \emph{Pacific Symposium on
  Biocomputing} \strong{13}:15--24.

Penn et al. 2010. An
  alignment confidence score capturing robustness to guide tree uncertainty.
  \emph{Molecular Biology and Evolution} \strong{27}:1759--1767.

Sela et al. 2015. GUIDANCE2: accurate detection of unreliable
  alignment regions accounting for the uncertainty of multiple parameters.
  \emph{Nucleic Acids Research} \strong{43}:W7--W14
}
\seealso{
\code{\link{guidance}}, \code{\link{HoT}}
}
\author{
Franz-Sebastian Krah
}
