% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidanceAA-class.R
\docType{class}
\name{guidanceAA-class}
\alias{guidanceAA-class}
\title{An S4 Class to represent Amino Acid Alignments with Quality Scores}
\description{
\code{"guidanceAA"} holds a multiple sequence alignment of alino
  acids together with quality scores for each cell in the alignment.
}
\section{Slots}{

\describe{
\item{\code{msa}}{An object of class \code{\link{AAbin}}.}

\item{\code{scores}}{A matrix of quality scores.}

\item{\code{method}}{A characters string giving the method to derive the quality
scores.}
}}

\seealso{
\code{"\link[=guidanceDNA-class]{guidanceDNA}"}
}
\author{
Christoph Heibl
}
