% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidanceAA.R
\name{guidanceAA}
\alias{guidanceAA}
\title{Create Objects of Class "guidanceAA"}
\usage{
guidanceAA(msa, scores, method, msa.method)
}
\arguments{
\item{msa}{An object of class \code{\link{AAbin}}.}

\item{scores}{A matrix of quality scores.}

\item{method}{A characters string giving the method to derive the quality
scores.}

\item{msa.method}{A character string giving the alignment method}
}
\description{
Create objects of class
  \code{"\link[=guidanceAA-class]{guidanceAA}"} from objects of class
  \code{"\link{AAbin}"} and a matrix derived from \code{\link{guidance}} or
  \code{\link{HoT}}.
}
\author{
Christoph Heibl
}
