% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidanceDNA-class.R
\docType{class}
\name{guidanceDNA-class}
\alias{guidanceDNA-class}
\title{An S4 Class to represent DNA Alignments with Quality Scores}
\description{
\code{guidanceDNA} holds a multiple sequence alignment of DNA
  together with quality scores for each cell in the alignment.
}
\section{Slots}{

\describe{
\item{\code{msa}}{An object of class \code{\link{DNAbin}}.}

\item{\code{scores}}{A matrix of quality scores.}

\item{\code{method}}{A characters string giving the method to derive the quality
scores.}
}}

\seealso{
\code{"\link[=guidanceAA-class]{guidanceAA}"}
}
\author{
Christoph Heibl
}
