% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidanceDNA.R
\name{guidanceDNA}
\alias{guidanceDNA}
\title{Create Objects of Class "guidanceDNA"}
\usage{
guidanceDNA(msa, scores, method)
}
\arguments{
\item{msa}{An object of class \code{\link{DNAbin}}.}

\item{scores}{A matrix of quality scores.}

\item{method}{A characters string giving the method to derive the quality
scores.}
}
\description{
Create objects of class
  \code{"\link[=guidanceDNA-class]{guidanceDNA}"} from objects of class
  \code{"\link{DNAbin}"} and a matrix derived from \code{\link{guidance}} or
  \code{\link{HoT}}.
}
\author{
Christoph Heibl
}
