% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidanceSA.R
\name{guidanceSA}
\alias{guidanceSA}
\title{Interface to guidance program}
\usage{
guidanceSA(seqs, bootstrap, msa.program, program, outorder, msafile,
  cutoff = 0.93, moreArgs, exec, proc_num, quiet = FALSE)
}
\arguments{
\item{seqs}{An object of class \code{\link{DNAbin}} or
\code{\link{AAbin}} containing unaligned sequences of DNA or amino acids.}

\item{bootstrap}{An integer giving the number of perturbated MSAs.}

\item{msa.program}{A charcter string giving the name of the MSA program,
currelty one of c("mafft", "muscle", "clustalo", "clustalw2"); MAFFT is
default}

\item{program}{A charcter string giving the name one of c("guidance", "guidance2", "HoT").}

\item{outorder}{As input or aligned. Default=aligned}

\item{msafile}{Path to user supplied local MSA file}

\item{cutoff}{Confidence cutoff between 0 to 1. Default=0.93}

\item{moreArgs}{More arguments passed to GUIDANCE}

\item{exec}{path to guidance program folder, e.g.
"/Applications/guidance.v2.02/"}

\item{proc_num}{Integer giving the number of cores.}

\item{quiet}{logical if TRUE, progress is printed to console.}
}
\value{
list containing following scores and alignments:

mean_scores residue pair score and mean column score

column_score

residue_column_score GUIDANCE score

residue_pair_residue_score

residual_pair_sequence_pair_score

residual_pair_sequence_score

residue_pair_score

base_msa

guidance_msa is the base_MSA removed from unreliable
  residues/columns/sequences below cutoffs
}
\description{
Interface to guidance program
}
\author{
Franz-Sebastian Krah
}
