% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoT.R
\name{HoT}
\alias{HoT}
\title{Heads or Tails Alignment Reliability}
\usage{
HoT(sequences, method = "auto", bootstrap,
  msa.exec = "/usr/local/bin/mafft", ncore = 1, zip.file)
}
\arguments{
\item{sequences}{object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing unaligned sequences of DNA or amino acids.}

\item{method}{further argument passed to MAFFT, default is \code{"auto"}}

\item{bootstrap}{integer giving the number of alternative MSAs to be computed}

\item{msa.exec}{A character string giving the path to the executable of the
alignment program (e.g. \code{/usr/local/bin/mafft}); possible programs are
\code{MAFFT}, \code{MUSCLE}, and \code{ClustalW}. 
For details see \code{\link{clustal}}, \code{\link{mafft}}}

\item{ncore}{integer specifying the number of cores; default = 1 (serial),
"auto" can be used for automated usage of all detected cores.}

\item{zip.file}{A character string giving the name for the output zip file.}
}
\value{
An object of class \code{\linkS4class{guidanceDNA}} or
  \code{\linkS4class{guidanceAA}}.
}
\description{
MSA reliability assessment HoT (Landan and Graur 2008)
}
\details{
Calculates column reliability (and other scors) by comparing
  alternative MSAs generated by aligning guide tree partitions as described
  in Landan and Graur (2008). For details see \code{compareMSAs}. 8*(N-3)
  alternative MSAs are generated by default, where N is the number of
  sequences.

For an example workflow see Vignette
}
\examples{
\dontrun{
# run GUIDANCE on example data using MAFFT
fpath <- system.file("extdata", "BB30015.fasta", package="rGUIDANCE") # random example from BALiBASE
fas <- ape::read.FASTA(fpath)
g <- HoT(sequences = fas, msa.exec= "/usr/local/bin/mafft")
scores <- scores(g, score = "column")
plot(scores$column$score, xlab = "Site", 
ylab = "Column score", 
main = "HoT", type ="l")
}

}
\references{
Landan and Graur. 2008. Local reliability measures from
  sets of co-optimal multiple sequence alignments. \emph{Pacific Symposium on
  Biocomputing} \strong{13}:15--24.
}
\seealso{
\code{\link{msa_set_score}}, \code{\link{guidance}},
  \code{\link{guidance2}}
}
\author{
Franz-Sebastian Krah
}
