% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoT_in_GUIDANCE2.R
\name{Hot_GUIDANCE2}
\alias{Hot_GUIDANCE2}
\title{HoT within guidance2}
\usage{
Hot_GUIDANCE2(msa, n.coopt, raw_seq, method, msa.exec)
}
\arguments{
\item{msa}{An object of class \code{\link{DNAbin}} or
\code{\link{AAbin}} containing aligned sequences of DNA or amino acids.}

\item{n.coopt}{An integer giving the number of sampled co-optimal MSAa.}

\item{raw_seq}{An object of class \code{\link{DNAbin}} or
\code{\link{AAbin}} containing unaligned sequences of DNA or amino acids.}

\item{method}{A character string containing further arguments passed to
MAFFT; default is \code{"auto"}.}

\item{msa.exec}{A character string giving the path to the executable of the
alignment program (e.g. \code{/usr/local/bin/mafft}); possible programs are
\code{MAFFT}, \code{MUSCLE}, and \code{ClustalW}.}
}
\description{
Internally Heads or Tails Alignment Reliability within guidance2
}
\author{
Franz-Sebastian Krah
}
