% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_set_scoreR.R
\name{msa_set_score}
\alias{msa_set_score}
\title{Compare reference MSA with alternative MSA(s)}
\usage{
msa_set_score(ref, alt)
}
\arguments{
\item{ref}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing \strong{aligned} sequences of DNA or amino acids.}

\item{alt}{single MSA or list of MSAs or path to alternative files. Single
MSAs and list members should be of class \code{\link{DNAbin}} or
\code{\link{AAbin}}.}
}
\value{
matrix containing following scores:

residue_pair_score: if one alternative MSA is supplied then the score
  is 1 if a residue pair was identically aligned as in the reference MSA and
  0 otherwise. If more than one alternative MSA is used, then the average of
  the residue pair scores from all comparisons
}
\description{
MSA reliability scores (Penn et al. 2010)
}
\details{
In principle this can also be done via AlignStat::compare_alignments, but slower.
}
\references{
Penn et al. 2010. An alignment confidence score capturing
  robustness to guide tree uncertainty. \emph{Molecular Biology and
  Evolution} \strong{27}:1759--1767.
}
\author{
Franz-Sebastian Krah
}
