% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaBP.R
\name{msaBP}
\alias{msaBP}
\title{Bootstrap a Multiple Sequence Alignnment}
\usage{
msaBP(msa)
}
\arguments{
\item{msa}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}.}
}
\description{
Create a bootstrap replicate of multiple sequence alignment.
}
\author{
Franz-Sebastian Krah
}
