% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaBP_nj_tree.R
\name{msaBP_nj_tree}
\alias{msaBP_nj_tree}
\title{Neighbor joining BP Tree from MSA}
\usage{
msaBP_nj_tree(msa, outgroup = "auto")
}
\arguments{
\item{msa}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing \strong{aligned} sequences of DNA or amino acids.}

\item{outgroup}{A character string giving the taxon name to be used for
rooting; the default (\code{"auto"}) chooses the first taxon.}
}
\value{
tree of class phylo
}
\description{
Estimate a neighbor joing tree from a multiple sequence alignment.
}
\author{
Franz-Sebastian Krah
}
