% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{partitions}
\alias{partitions}
\title{Create N-3 Partitions from Phylogeny}
\usage{
partitions(tree)
}
\arguments{
\item{tree}{A phylogenetic tree of class \code{\link{phylo}}.}
}
\description{
A wrapper to \code{\link{treePart}} from the \strong{adephylo}
  package.
}
\author{
Franz-Sebastian Krah
}
