% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_recode.R
\name{recodeMSA}
\alias{recodeMSA}
\title{Cmatrix calculation (for developers)}
\usage{
recodeMSA(msa)
}
\arguments{
\item{msa}{A MSA of class \code{\link{DNAbin}} or \code{\link{AAbin}}}
}
\value{
a list with integers
}
\description{
Meant for developers only.
}
\details{
Here implemented as in GUIDANCE program msa_set_score. Characters are represented by odd numbers 
and gaps by even numbers. 
For more info about the Cmatrix recoding see page 3 on the supplemantary material of: Satija et al. (2009)
http://www.biomedcentral.com/content/supplementary/1471-2148-9-217-s1.pdf

For an example see \code{\link{addMSA}}
}
\references{
Satija R, Novak A., Mikls I., Lyngs R., and Hein J. (2009) BigFoot: 
Bayesian alignment and phylogenetic footprinting with MCMC, 
\emph{BMC Evolutionary Biology} \strong{9}:217

Penn et al. 2010. An
  alignment confidence score capturing robustness to guide tree uncertainty.
  \emph{Molecular Biology and Evolution} \strong{27}:1759--1767.
}
\author{
Franz-Sebastian Krah
}
