% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_DNA.R
\name{rev_DNA}
\alias{rev_DNA}
\title{Reverse DNA Sequences of class DNAbin or AAbin}
\usage{
rev_DNA(x)
}
\arguments{
\item{x}{An object of class \code{\link[ape]{DNAbin}} or \code{\link[ape]{AAbin}}.}
}
\description{
Creates the reverse of DNA of amino acid sequences.
}
