% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGUIDANCE-package.R
\docType{package}
\name{rGUIDANCE}
\alias{rGUIDANCE}
\alias{rGUIDANCE-package}
\title{rGUIDANCE: GUIDe tree based AligNment ConfidencE (GUIDANCE) in R}
\description{
rGUIDANCE: GUIDe tree based AligNment ConfidencE (GUIDANCE) in R
}
