% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMSA.R
\name{filterMSA}
\alias{filterMSA}
\title{Filtering of the base MSA using guidance scores}
\usage{
filterMSA(guidanceX, col.cutoff = 0.2, seq.cutoff = 0.1,
  mask.cutoff = 0.5, filter.ends = FALSE, filter.gaps = FALSE,
  column_score = FALSE, flag_a = FALSE, na.coding = 0.5)
}
\arguments{
\item{guidanceX}{object of class \code{\link{guidance}}}

\item{col.cutoff}{numeric between 0 and 1; removes unreliable columns below
the cutoff (default: 0.2); ignored if FALSE}

\item{seq.cutoff}{numeric between 0 and 1; removes unreliable sequences below
the cutoff (default: 0.1); ignored if FALSE}

\item{mask.cutoff}{residues below the cutoff are masked ('N' for DNA, 'X' for
AA; default: 0.5); ignored if FALSE}

\item{filter.ends}{logical, if TRUE trim.ends (ips) is applied to the MSA}

\item{filter.gaps}{logical, if TRUE trim.gabs (ips) is applied to the MSA}

\item{column_score}{logical, if TRUE column score (e.g. for RAxML: flag -a)
is in the output}

\item{flag_a}{character specifying a path. If path is supplied function
writes the filtered MSA into a fasta file. Additionally the function
produces a file with the column score ready for RAxML input (flag -a)}

\item{na.coding}{value with which to replace NAs}
}
\value{
masked MSA of class \code{AAbin} or \code{DNAbin}

column_score is optional
}
\description{
Filtering of the base MSA using guidance scores
}
\seealso{
\code{\link{scores}}
}
\author{
Franz-Sebastian Krah
}
